refactor(catridge): modularize
